messagesTemplate.json will be created on build. And all generated js & json files got moved to jsBuilds
