+++ .gitignore
@@ -26,6 +26,7 @@ build/Release
 # https://www.npmjs.org/doc/misc/npm-faq.html#should-i-check-my-node_modules-folder-into-git
 node_modules
 
-# Client code generated by browserify
-bundle.js
+# Client code generated by browserify and other JS- and JSON-files generated
+jsBuilds/
+!jsBuilds/info.txt
 *.map
