+++ tools/master_message_template.msg
